Move preview to it's own file
