+++ src/main/webui/src/ProposalEditorView/proposal/Overview.tsx
@@ -6,7 +6,6 @@ import {
 import {
     Accordion,
     Avatar,
-    Badge,
     Box,
     Container,
     Group,
@@ -18,7 +17,7 @@ import {
     CalibrationObservation,
     CalibrationTargetIntendedUse,
     Investigator,
-    RealQuantity, TextFormats
+    RealQuantity,
 } from 'src/generated/proposalToolSchemas.ts';
 import { IconNorthStar } from '@tabler/icons-react';
 import { ReactElement, useRef } from 'react';
@@ -27,9 +26,7 @@ import downloadProposal from './downloadProposal.tsx';
 import { DIMMED_FONT_WEIGHT, JSON_SPACES } from 'src/constants.tsx';
 import { TargetTable } from '../targets/TargetTable.tsx';
 import { TechnicalGoalsTable } from '../technicalGoals/technicalGoalTable.tsx';
-import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import Latex from "react-latex-next";
-import "katex/dist/katex.min.css";
+import {PreviewJustification} from "../justifications/justification.preview.tsx";
 
 /*
       title    -- string
@@ -59,48 +56,6 @@ import "katex/dist/katex.min.css";
         of the overview plus the supporting document files
  */
 
-/**
- * Displays a preview of a justification, either with syntax highlighting or rendered preview
- *
- * @param {TextFormats} format enum of possible formats (asciitext, latex etc.)
- * @param {string} content the actual justification text to be previewed
- * @constructor
- */
-export const PreviewJustification = (format: TextFormats, content: string) => {
-    switch (format) {
-        case 'asciidoc':
-            return (<>
-                <Badge>{format}</Badge>
-                <Box bg={"gray.2"} c={"black"} p={"md"} m={"xs"}>
-                    <SyntaxHighlighter language={'asciidoc'}>
-                        {content}
-                    </SyntaxHighlighter>
-                </Box>
-            </>);
-
-        case 'latex':
-            return (
-                <>
-                    <Badge>{format}</Badge>
-                    <Box bg={"gray.2"} c={"black"} p={"md"} m={"xs"}>
-                        <Latex>
-                            {content}
-                        </Latex>
-                    </Box>
-                </>);
-
-        default:
-            return (
-                <>
-                    <Badge>{format}</Badge>
-                    <Box bg={"gray.2"} c={"black"} p={"md"} m={"xs"}>
-                        <SyntaxHighlighter language={'asciidoc'}>
-                            {content}
-                        </SyntaxHighlighter>
-                    </Box>
-                </>);
-    }
-}
 
 
 /**
